fix: enum not found in DatePicker and SwipeAction components
