+++ packages/bui-core/src/DatePicker/constants.ts
@@ -1,0 +1,10 @@
+export enum DatePickerType {
+  YEAR = 'year',
+  MONTH = 'month',
+  DAY = 'day',
+  HOUR = 'hour',
+  MINUTE = 'minute',
+  SECOND = 'second',
+}
+
+export default DatePickerType;
