+++ packages/bui-core/src/SwipeAction/constants.ts
@@ -1,0 +1,10 @@
+export enum SideTypeEnum {
+  LEFT = 'left',
+  RIGHT = 'right',
+}
+
+export enum DragPhaseEnum {
+  START = 1,
+  MOVE = 2,
+  END = 3,
+}
